Factory Pattern applied to GameObject Creation
In the process of making the gameobject render recursively, I came across the fact that I would have to grab the transform and the renderer every frame that we want to render.
The system that I came up with along with Copilot was a for loop that did a dynamic cast, which checked if it was the same as the given type, if not it would return a nullptr. However this is a pretty heavy check, and therefore its better to cache the transform and renderer when they are added to the GameObject.

This is now done through use of a Factory Pattern that creates the object and adds a transform component on init. This can later be extended to also include the Renderer (though that depends on if its a empty object.

Still not fully finished the recursive rendering however.

#2 #10 #4