Fixed camera input
Problem lies with ImGui being a bit bugged with docked spaces and wanting the keyboard if a panel is selected, even if it doesn't actually need the user input. By removing the second if statement we avoid that, as we only want to move around if the viewport is focused anyways.

#closes 24