Implemented Directional and Spot lights
Changed OpenGL version to `4.3` instead of `3.3`. Updated `lighting.glsl` to handle the two different types of lights.
Tried converting the UBO to SSBO, but that was a big ass mess. Main issue that I ran into was being unable to get the data from the CPU to the GPU. Due to one issue or another, the data never transferred correctly, so no lights were registered (GPU side). Gave up on that and just made use of the UBO, as its working fine, it just needs a bit of padding depending on the field types and array sizes.

Changelog:

- `.frag` and `.vert`:
  - #version 430 instead of #version 400.
- `lighting.glsl`:
  - Now using a switch case in the for-loop for the lightType. applies the light effect differently depending on the light type.
  - Added `vec4 directions[4]` and `ivec4 lightTypes[4]` to the uniform LightBlock UBO.

- Light:
  - Added `ToInt()` inline helper function to cast the lightType to a int.
  - Updated the `LightData` struct to support lightType and directions.

- Transform:
  - Added `forward()`. Returns the forward vector of the gameObject.

- Scene:
  - Updated the Render loop to get and send the new lightData to the GPU.
  - removed `mat4 identity` as it was not being used.

- main:
  - Version_Major changed from 3 to 4.
  - Replaced the ungodly amount of UniformBlockBinding calls with `glBindBuffer (GL_UNIFORM_BUFFER, 0);`. This is possible due to the OpenGL version upgrade.

closes #31
closes #30
closes #28