Created `FrameBuffer` class to wrap the usage of FBO's
Added a FrameBuffer class to wrap the usage of different FBO types. This was done within the specifications as shown in #38. A few things different compared to the issue example is the `enum class AttachmentType`, which had unclear values. These have been changed to GL safe values. `FrameBufferSpecifications` has also been changed to make use of the `AttachmentTypes` enum. FrameBuffer class itself has more methods—both public as well as private—then in the example. This was to keep the code clean, as well as to be able to handle destruction (making use of a destructor as well as a `Destroy()` method), and different attachment types. The class has getters to allow you to do comparisons outside of the `FrameBuffer` class.

Changelog:

`frameBuffer`:
- Add FrameBuffer class encapsulating OpenGL FBO creation, resizing, binding, and destruction.
- Support multiple attachment types (color, depth, stencil) by making use of a switch case that calls different combinations of methods:
  - `AttachColor(const int w, const int h)`. -> most of the methods have `const` parameters, to ensure they are not changed arbitrarily along the process of FBO creation.
  - `AttachDepth(const int w, const int h)`.
  - `AttachDepthStencil(const int w, const int h)`.
  - `AttachDepthTexture(const int w, const int h)`.

`viewportPanel`:
- Refactor ViewportPanel to use FrameBuffer class.
- Remove manual FBO, texture, and renderbuffer management.
- Update draw method to use FrameBuffer API for resizing and rendering.

`editor.cpp`:
- Update viewport width/height accessors to use FrameBuffer getters.

`material.h`:
- Set default value for `TextureData slot` to 0.

`CMakeLists.txt`:
- Add frameBuffer.cpp to build sources