Implemented a Post-Processing Effect Base Class
Created `class PostProcessingEffectBase` to act as a base class for Post-Processing effects. Opted to go for a base class as its more readable when dealing with multiple effects at the same time, as well as keeping `DrawGui()` unique to each effect type contained to their own file.
Added `postProcessingEffectBase.cpp` to `CMakeLists.txt`, and moved the uniform set for bools in `material.cpp` between the int and vec3 loop to represent the order of maps in `material.h`.

Changelog:
- `postProcessingEffectBase`:
  - Constructor that required name and material to be passed.
  - Getter an setter for enabled state.
  - Getter for name and Material.
  - virtual methods for Applying the effect as well as drawing to the GUI.
    - `Apply(GLuint inputTexture, FrameBuffer& output, int width, int height)` -> basic implementation where the effect is applied once. Effects that need to apply it multiple times need to override the method and call base multiple times, or have their own implementation.

- `Material.cpp`
  - Moved for-loop to set the bool uniforms up in the order, in between the int and vec3 for loops. Now the order is the same as the maps are ordered in the header file.

- `CMakeLists.txt`
  - Added postProcessingEffectBase to the library list.

closes #39
#10 #37