Fixed #54 - Turning off all PP effects freezes viewport
The issue indeed was with the edge-case handling in `PostProcessingManager::ProcessStack()`.
Added `int skippedEffects` to keep track of the amount of effects skipped, and is later evaluated to see if all effects in the stack are skipped. The evaluation is checking if the amount of skipped effects is as big as the list of effects in the stack, if that is the case, we blit the inputBuffer to the outputBuffer.

Changelog:

- global:
  - Fixed `postProcessingEffectBase->isEnabled` calls to make use of the `IsEnabled()` getter function.
  - Fixed `postProcessingEffectBase->isEnabled = bool` calls to make use of the `SetEnabled()` setter function.

- `postProcessingManager`:
  - Added no effects applied case handling. Now copies inputBuffer to outputBuffer if no effects are applied.
  - Added printf statement for debugging purposes.

closes #54
closes #40
#37