Major overhaul of post processing effects
Overall pretty big overhaul of how I am applying post-processing effects.
The `PostProcessingManager` had a fundamental redesign, instead of iterating over effects in the m_effects list, and first doing a check if they are enabled or not, it now keeps track of another list that holds the enabled effects (`m_enabledEffects`). This allow us to keep better track of what effect is the last effect, without having to do a initial check to see which effect is the last actual activated effect. It now created the effects itself in its Initialization method, ought to be called after the manager is created, there it creates the effects and puts them in the list. The manager now also holds only one temp framebuffer instead of 2. This allows us to let the effects that actually need to ping pong to handle that themselves, and keeps the responsibility of the manager clear.

`PostProcessingBaseEffect` has been made to be more robust/full on its own, meaning that effect implementations need to create less of their own fields and can use the base fields. The `Apply()` method signature has shifted from raw texture ID's to framebuffer wrapper references. This provides more type safety and encapsulation, access to the framebuffer metadata (name, is_valid, etc) which can be used for debugging purposes or more safety checks, and leads to a more consistent signature—input and output are the same type, instead of different ones. `IsEnabled()` and `SetEnabled()` have been refactored into a `isEnabled` property, which has its SetOnChange set to call EnableEffect() and DisableEffect() on the manager if enabled or disabled. Quad rendering is now also handled by the PostProcessingEffect, instead of by the Manager.

`BloomEffect` has had a somewhat complete rewrite in its Apply(). Initially the manager had to call the apply multiple times, passing the index of the apply, now its done internally. It initially combined blur and composite into the same shader/material, and made use of a stored texture in a member variable (state leak). Now it makes use of a single entry architecture: The manager calls `Apply()` once, and the method goes into a for-loop for the multiple passes. Threshold -> Blur -> Composite is the new pipeline, with composite getting its own shader/material. The effect now holds its own ping-pong buffers (`tempFBO_1` and `tempFBO_2`). Finally the debug modes have been implemented. While threshold works, blur only does not. This is due to the same issue that the effect as a whole doesn't work.

`InvertEffect` is a basic effect that inverts the colours, so black becomes white, red becomes a green/blue colour, etc. This works as meant, and mainly makes use of the base class applications, only really setting up the material and shader used. The shader itself is very simple: `Sample texture -> do 1 - sampledColour -> tadah`.

`FrameBuffer`'s have also received a upgrade! Framebuffers now have names for debugging purposes. This allows you to check what fbo is being bound at a certain moment for example. Validation and error reporting have been added.

Documentation has been added to materials and PostProcessingManager. `.gitignore` has been updated to exclude captures folder, for RenderDoc Captures. CMake has been updated to add the new files (invertEffects.cpp and postProcessingEffects.h). Added a openGL Debug Context as a callback when openGL errors are thrown. PostProcessPanel now makes use of the isEnabled property bool.

Changelog:

- post-processing shaders:
  - Added `composite.frag`
  - Added `invert.frag`

- `bloomEffect`:
  - Added `m_compositeMaterial` field, as well as fields for the materials associated shaders.
    - Updated Constructor to create the shaders and create the materials from the shaders, storing both in the process.
  - Creation and storing of 2 ping-pong buffers.
    - Updated Constructor to create these buffers.
  - Changed `Apply()`:
    - Changed signature to correspond with postProcessingEffectsBase's `Apply()` signature.
    - Does first pass, then for loop for blur passes, then composite pass. Makes use of the passCount only for the loop.
  - Doesn't yet fully work ;-;

- `invertEffect`:
  - Basic invert of colours effect.

- `postProcessingEffects`:
  - Encapsulating #include header, used in `postProcessingManager` to get all the post-processing effects in one go.

- `postProcessingEffectBase`:
  - Refactored enabled state to make use of the property type.
  - Refactored Apply():
    - Changed the signature to use just pure FBO wrappers.
    - Added shader property to be used by inheriting classes.
    - Replaced glClear and glViewport with Macro CLEAR_BOUND() to allow for debugging.
    - Added RenderQuad call to the base implementation.
  - Added RenderQuad() to be used by base implementation as well as inheriting effects.

- `postProcessingManager`:
  - Added documentation to all properties and methods, as well as the class itself.
  - Removed the `RemoveEffect()` method as it won't be used. It has been repalced with a `DisableEffect()` method instead.
  - Added EnableEffect and Initialize methods().
    - Enable/Disable Effect do as you would guess. Enable sorts the m_enabledEffects vector.
  - Initialize creates the post-processing effects.
  - Creating and storing just one FBO now.
  - Another vector of enabled effects.
  - Redid ProcessStack to make use of the new single-entry architecture of post-processing effects.

- `frameBuffer`:
  - Added a name field for debugging.
  - Bind() now checks the validity and logs a error if invalid.
  - BindRead has been added.
  - ClearBound method has been added with additional macro for file and line debugging.

- `material`:
  - Added documentation

- Additional:
  - Updated CMakeList.txt for new files
  - Updated main with new openGLDebug callback.
  - Added captures folder to .gitignore.
  - Viewport now makes use of the isEnabled property<bool> of postProcessingEffectBase.

commit 089424d8105b4ed9376cc96d5762874c4b369232
Author: Beer Huizer <beer.huizer@gmail.com>
Date:   Wed Dec 17 22:23:35 2025 +0100

    Sort of working post processing

    Manager now seems to fully work, and so does the invert, its bloom thats being a bitch.

commit 16f11fbd74daba5035fe72f6e9e13b6c41f8a080
Author: Beer Huizer <beer.huizer@gmail.com>
Date:   Wed Dec 17 19:31:17 2025 +0100

    Added working Invert postProcessing Effect

    Now rework the bloom effect.

commit b68ec965d6b3547e86950a33f665e6ba12e6aa52
Author: Beer Huizer <beer.huizer@gmail.com>
Date:   Sun Dec 14 22:18:18 2025 +0100

    Attempt at rework with lucas

    We ballin