Added light intensity support
Added light intensity support by making it the .w component of the vec4 color. This means that we don't have to make another field in the UBO and keep the packing the same. In lighting.glsl we now use read the colors[i].w value to get the intensity of a light. This then gets added to the diffuse values.

Changelogs:
- `light`:
  - Added `UpdateRendererIntensity()`.
  - Made `Property<float> intensity` actually have a callback -> `UpdateRendererIntensity()`.

- `scene`:
  - Now implant the light intensity into the .w component of the light color.
  - Removed redundant printf statements as well as comments.
  - Made many if-statements inline.

- `shaders`:
  - `Lighting.glsl` now uses the .w component of the vec4 color to get the intensity of the light.
  - `fragmentLightBulb.frag` now multiplies the color with the intensity, making it go into HDR value faster for bloom effect to apply.
  - Redid the normal in `litFragment.frag` calculation to default to the normalized value of fNor, as well as removing redundant lines.